🐛 Fix HdTimeslots story exports to work with the new version
Storybooks uses Component Story Format (CSF), if we export something that don't follow this format, the application crashes.

I extracted the function into a file to prevent this error.
